+++ fastapi_app/src/infrastructure/repos/post_rep.py
@@ -1,10 +1,14 @@
 from typing import Type, List, Optional
 from sqlalchemy.orm import Session, joinedload
 from datetime import datetime
+from sqlalchemy.exc import IntegrityError
 
 from infrastructure.models.posts_model import Post
 from infrastructure.models.categories_model import Category
+from infrastructure.models.locations_model import Location
+from infrastructure.models.users_model import User
 from schemas.posts import PostRequest, PostUpdate
+from core.exceptions.infrastructure_exceptions import *
 
 
 class PostRepository:
@@ -13,6 +17,7 @@ class PostRepository:
 
     def get_posts(self, session: Session, skip: int = 0, limit: int = 20) -> List[Post]:
         now = datetime.now()
+
         query = (
             session.query(self._model)
             .outerjoin(self._model.category)
@@ -33,7 +38,10 @@ class PostRepository:
             .options(joinedload(self._model.category))
             .where(self._model.id == post_id)
         )
-        return query.scalar()
+        post = query.scalar()
+        if not post:
+            raise PostNotFoundById
+        return post
 
     def get_by_author(
         self, session: Session, author_id: int, skip: int = 0, limit: int = 20
@@ -45,6 +53,9 @@ class PostRepository:
             .offset(skip)
             .limit(limit)
         )
+        author = session.query(User).where(User.id == author_id).scalar()
+        if not author:
+            raise UserDoesNotExist
         return query.all()
 
     def get_by_category(
@@ -57,29 +68,65 @@ class PostRepository:
             .offset(skip)
             .limit(limit)
         )
+        category = session.query(Category).where(Category.id == category_id).scalar()
+        if not category:
+            raise CategoryNotFoundById
         return query.all()
 
     def create_post(self, session: Session, data: PostRequest, author_id: int) -> Post:
-        new_post = self._model(
-            **data.model_dump(),
-            author_id=author_id,
-            pub_date=datetime.now(),
-            created_at=datetime.now(),
-        )
-        session.add(new_post)
-        session.commit()
-        session.refresh(new_post)
-        return new_post
+        try:
+            author = session.query(User).where(User.id == author_id).scalar()
+            if not author:
+                raise UserNotFoundById
+            if data.category_id:
+                category = session.query(Category).where(Category.id == data.category_name).scalar()
+                if not category:
+                    raise CategoryNotFoundByName
+                if not category.is_published:
+                    raise CategoryNotPublished
+
+            if data.location_id:
+                location = session.query(Location).where(Location.id == data.location_id).scalar()
+                if not location:
+                    raise LocationNotFoundById
+                
+        
+            new_post = self._model(
+                **data.model_dump(),
+                author_id=author_id,
+                pub_date=datetime.now(),
+                created_at=datetime.now(),
+            )
+            session.add(new_post)
+            session.flush()
+            session.refresh(new_post)
+            return new_post
+        
+        except IntegrityError as e:
+            session.rollback()
+            raise DatabaseIntegrityError(
+                original_error=e
+            )
 
     def update_post(self, session: Session, post: Post, data: PostUpdate) -> Post:
         up_post = data.model_dump(exclude_unset=True)
         for key, value in up_post.items():
             setattr(post, key, value)
-
-        session.commit()
-        session.refresh(post)
+        exist = self.get_by_id(session, post.id)
+        if not exist:
+            raise PostDoesNotExist
+        if data.location:
+            location = session.query(Location).where(Location.name == data.location_name).scalar()
+            if not location:
+                raise LocationNotFoundByName
+        if data.category:
+            category = session.query(Category).where(Category.slug == data.category_name).scalar()
+            if not category:
+                raise CategoryNotFoundByName
         return post
 
     def delete_post(self, session: Session, post: Post) -> None:
+        exist = self.get_by_id(session, post.id)
+        if not exist:
+            raise PostDoesNotExist
         session.delete(post)
-        session.commit()
